Move github_validators module into `lib`
